Bug 1181481 - Bring our patches from b2g-kitkat to b2g-studio-1.3. r=tdz

+++ android/main-emulator.c
@@ -34,7 +34,7 @@
 #include <android/avd/util.h>
 
 /* Required by android/utils/debug.h */
-int android_verbose;
+unsigned long long android_verbose;
 
 
 #define DEBUG 1
